requirements.txt

streamlit>=1.20.0
numpy>=1.23.0
pillow>=9.0.0
tensorflow>=2.10.0,<2.16.0
h5py>=3.7.0